+++ server/infra/mysql/music_dal.go
@@ -28,7 +28,7 @@ func NewMusicDal(db *gorm.DB) MusicDal {
 }
 
 type Music struct {
-	Id          int    `json:"id" gorm:"primaryKey;autoIncrement"`
+	Id          int64  `json:"id" gorm:"primaryKey;autoIncrement"`
 	RootId      int    `json:"root_id"`
 	Title       string `json:"title" gorm:"not null;uniqueIndex:uni_idx_music"`
 	ArtistIds   string `json:"artist_ids" gorm:"not null;uniqueIndex:uni_idx_music"`
